requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
langchain
langchain-openai
langchain-community
langgraph
chromadb
openai
tiktoken
python-multipart
python-dotenv
asyncio
aiohttp
streamlit
pandas
tabulate
pytesseract
pdf2image
pypdf
langchain-chroma
SQLAlchemy[asyncio]
aiosqlite
rank_bm25
